+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 Python PEP8 Autoformat
 
+## 2012.11.02-1
+ - Update autopep8 to version 0.8.2. This version fix [W601][fix-github-issue40]
+ - Add new setting "max-line-length" (default is 79).
+
 ## 2012.10.14-1
 - FIX: [Issue #4][issue4] (format only one round).
 - Upgrade shipped autopep8.